+++ sterrenboom/__manifest__.py
@@ -1,6 +1,6 @@
 {
     'name': 'Sterrenboom Parent Committee',
-    'version': '19.0.4.3.0',
+    'version': '19.0.4.4.0',
     'category': 'Services/Sterrenboom',
     'summary': 'Manage parent committee members and events for VBS De Sterrenboom',
     'description': """
@@ -41,6 +41,7 @@ Tools for managing the parent committee (oudercomité) of VBS De Sterrenboom:
         'views/sale_order_views.xml',
         'views/event_templates.xml',
         'views/event_halloweentocht_templates.xml',
+        'views/branding_templates.xml',
         'data/mail_template_data.xml',
         'data/event_halloweentocht_data.xml',
     ],