+++ sterrenboom/models/__init__.py
@@ -1,6 +1,7 @@
 from . import account_move
 from . import event_event
 from . import event_registration
+from . import mail_mail
 from . import sale_order
 from . import sterrenboom_event
 from . import sterrenboom_member