+++ sterrenboom/tests/__init__.py
@@ -1,3 +1,4 @@
+from . import test_branding
 from . import test_event_halloweentocht
 from . import test_event_registration_payment
 from . import test_mail_identity